fix(structure): move preview to it's own file, fix title
