+++ packages/sanity/src/structure/panes/documentList/sheetList/useDocumentSheetColumns.tsx
@@ -1,102 +1,20 @@
-import {ErrorOutlineIcon} from '@sanity/icons'
 import {isObjectSchemaType, type ObjectSchemaType} from '@sanity/types'
-import {Box, Checkbox, Flex, Stack, Text, Tooltip} from '@sanity/ui'
+import {Checkbox, Flex} from '@sanity/ui'
 import {
   type AccessorKeyColumnDef,
-  type CellContext,
   createColumnHelper,
   type GroupColumnDef,
   type VisibilityState,
 } from '@tanstack/react-table'
 import {useMemo} from 'react'
-import {useObservable} from 'react-rx'
-import {type DocumentPreviewStore, getPreviewStateObservable, useDocumentPreviewStore} from 'sanity'
+import {useDocumentPreviewStore} from 'sanity'
 
-import {useValidationMarkers} from '../../../../core/form/studio/contexts/Validation'
-import {PaneItem} from '../../../components'
-import {ValidationCard} from '../../document/inspectors/validation/ValidationInspector'
 import {DocumentSheetListSelect} from './DocumentSheetListSelect'
+import {PreviewCell} from './DocumentSheetPreviewCell'
 import {SheetListCellInner} from './SheetListCell'
 import {type DocumentSheetTableRow} from './types'
 
 export const VISIBLE_COLUMN_LIMIT = 5
-
-const RowValidation = ({
-  value,
-  schemaType,
-}: {
-  value: DocumentSheetTableRow
-  schemaType: ObjectSchemaType
-}) => {
-  const validation = useValidationMarkers()
-  if (validation.length === 0) {
-    return null
-  }
-  return (
-    <Tooltip
-      portal
-      content={
-        <Stack space={2} paddingY={2}>
-          {validation.map((marker) => (
-            <ValidationCard
-              key={marker.path.join('.')}
-              marker={marker}
-              schemaType={schemaType}
-              value={value}
-            />
-          ))}
-        </Stack>
-      }
-    >
-      <ErrorOutlineIcon />
-    </Tooltip>
-  )
-}
-interface PreviewCellProps extends CellContext<DocumentSheetTableRow, unknown> {
-  documentPreviewStore: DocumentPreviewStore
-  schemaType: ObjectSchemaType
-}
-
-const PreviewCell = (props: PreviewCellProps) => {
-  const paneProps = props.table.options.meta?.paneProps
-
-  const {documentPreviewStore, row, schemaType} = props
-  const title = 'Document title'
-  const previewStateObservable = useMemo(
-    () => getPreviewStateObservable(documentPreviewStore, schemaType, row.original._id, title),
-    [documentPreviewStore, row.original._id, schemaType],
-  )
-  const {draft, published, isLoading} = useObservable(previewStateObservable, {
-    draft: null,
-    isLoading: true,
-    published: null,
-  })
-  if (isLoading) {
-    return (
-      <Text size={1} muted>
-        Loading...
-      </Text>
-    )
-  }
-  const displayValue = (draft?.title ?? published?.title ?? 'Untitled') as string
-  const id = row.original.__metadata.idPair.publishedId
-  const isSelected = paneProps?.isActive && paneProps.childItemId === id
-
-  return (
-    <Flex align="center" gap={3} flex={1} paddingX={2}>
-      <Box flex={1}>
-        <PaneItem
-          schemaType={schemaType}
-          value={row.original}
-          title={displayValue}
-          id={id}
-          selected={isSelected}
-        />
-      </Box>
-      <RowValidation schemaType={schemaType} value={row.original} />
-    </Flex>
-  )
-}
 
 const columnHelper = createColumnHelper<DocumentSheetTableRow>()
 const SUPPORTED_FIELDS = ['string', 'number', 'boolean']
